COMP: Fix VS 17.10 compatibility with older runtime
Slicer CLI executables use the system location MSVC runtime instead of the one pre-packaged in the Slicer /bin directory. This resulted in an incompatibility of the application built with a newer version of the runtime than what was being used. This caused Slicer CLI executables such as ResampleScalarVolume to terminate with a fault.

Co-authored-by: Andras Lasso <lasso@queensu.ca>